\name{asuR-package}
\alias{asuR-package}
\alias{asuR}
\docType{package}
\title{
Advanced statistics using R
}
\description{
Functions and data sets used for a statistic course using R.
}
\details{
\tabular{ll}{
Package: \tab asuR\cr
Type: \tab Package\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{thomas.fabbro@unibas.ch
}
\keyword{ package }

