\name{dep-methods}
\docType{methods}
\alias{dep-methods}
\alias{dep,ANY-method}
\alias{dep,glm-method}
\title{Methods for Function dep in Package `asuR'}
\description{
Creates a plot of deviance residuals vs. linear predictor (eta).
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{Generic function}

\item{mymodel = "glm"}{Generates a plot of an object of class "glm"}
}}
\keyword{methods}
