\name{grasshopper}
\alias{grasshopper}
\alias{grasshoppers}
\docType{data}
\title{Femur length of grasshoppers}
\description{
Body size of Omocestus viridulus grasshoppers sampled in 13 populations at different altitudes in Switzerland. Data are from females and represent hind femur length in mm.
}
\usage{data(grasshopper)}
\format{
  A data frame with 205 observations on the following 3 variables.
  \describe{
    \item{\code{location}}{a factor with levels \code{baeretswil} \code{bendel} \code{birchwil} \code{elm} \code{flums} \code{gamserrugg} \code{hoernli} \code{naefels} \code{neerach} \code{pizol} \code{rothenthurm} \code{schoenenberg} \code{speer}}
    \item{\code{altitude}}{a numeric vector, meter above sea level}
    \item{\code{bodySize}}{a numeric vector, femur length in mm}
  }
}
\details{
For details see Berner Blanckenhorn (2006), Journal of Animal Ecology, 75:130?139.
}
\examples{
data(grasshopper)
}
\keyword{datasets}
