\name{hnp-methods}
\docType{methods}
\alias{hnp-methods}
\alias{hnp,ANY-method}
\alias{hnp,glm-method}
\title{ Methods for Function hnp in Package `asuR' }
\description{
 Methods for function \code{hnp} in Package `asuR'
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{generic function}

\item{mymodel = "glm"}{Diagnostics plots for objects of class "glm"}
}}
\keyword{methods}
