\name{houseflies}
\alias{houseflies}
\docType{data}
\title{Housfly Development}
\description{
Developmental duration of three housefly strains for two different treatments.
}
\usage{data(houseflies)}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{\code{duration}}{a numeric vector}
    \item{\code{treatment}}{a factor with levels \code{contr} \code{treat}}
    \item{\code{strain}}{a factor with levels \code{BE} \code{BW} \code{OL}}
  }
}

\source{
thomas.fabbro@unibas.ch
}
\examples{
data(houseflies)
}
\keyword{datasets}
