\name{identifyControl}
\alias{identifyControl}
\title{Takes control over the point identification}
\description{
internal use only
}
\usage{
identifyControl(panel.matrix, original.row.names, id, slot.names = NULL)
}
\arguments{
  \item{panel.matrix}{has the same structure as \code{\link{trellis.currentLayout}}}
  \item{original.row.names}{asks for the row.names in the original data.frame}
  \item{id}{see \code{\link{inspect}}}
  \item{slot.names}{not yet...}
}
\details{
}
\value{
an invisible vector  with the row number of the identified points
}
\author{\email{thomas.fabbro@unibas.ch}}
\note{under construction
}
\seealso{\code{\link{inspect}}}
\examples{
##
}
\keyword{internal}
