\name{inspect-methods}
\docType{methods}
\alias{inspect-methods}
\alias{inspect,ANY-method}
\alias{inspect,lm-method}
\alias{inspect,glm-method}
\alias{inspect,lmer-method}
\title{Methods for Model Inspection in Package `asuR'}
\description{
 Inspection of model objects using diagnostic plots
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{Generic funciton}

\item{mymodel = "lm"}{Diagnostic plots for objects of class "lm"}

\item{mymodel = "glm"}{Diagnostics plots for objects of class "glm"}

\item{mymodel = "lmer"}{Diagnostics plots for objects of class "lmer"}
}}
\keyword{methods}

