\name{length-methods}
\docType{methods}
\alias{length-methods}
\alias{length,ANY-method}
\alias{length,pDescList-method}
\title{Length of a pDescList-object}
\description{
Returns the length of a pDescList-object
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{ ~~describe this method here }

\item{\code{signature(x = "pDescList")}}{ ~~describe this method here }
}}
\keyword{methods}

