\name{lep-methods}
\docType{methods}
\alias{lep-methods}
\alias{lep,ANY-method}
\alias{lep,glm-method}
\title{ Methods for Function lep in Package `asuR' }
\description{
 Methods for function \code{lep} in Package `asuR'
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{generic function}

\item{mymodel = "glm"}{Diagnostics plots for objects of class "glm"}
}}
\keyword{methods}
