\name{mytrees}
\alias{mytrees}
\docType{data}
\title{Simulated tree data}
\description{
Similar to the data frame called trees in the dataset package. But
simulated data for learning regressions.
}
\usage{data(mytrees)}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{Volume}}{a numeric vector}
    \item{\code{Girth}}{a numeric vector}
    \item{\code{Height}}{a numeric vector}
  }
}
\examples{
data(mytrees)
}
\keyword{datasets}
