\name{nep-methods}
\docType{methods}
\alias{nep-methods}
\alias{nep,ANY-method}
\alias{nep,lmer-method}
\title{Methods for Function nep in Package `asuR'}
\description{
 Methods for function \code{nep} in Package `asuR'
}
\section{Methods}{
\describe{
\item{mymodel = "ANY"}{ generic function }

\item{mymodel = "lmer"}{ see the documentation of nep.lmer} 
}}
\keyword{methods}
