\name{nrp-methods}
\docType{methods}
\alias{nrp-methods}
\alias{nrp,ANY-method}
\alias{nrp,lmer-method}
\title{Methods for Function nrp in Package `asuR'}
\description{
 Methods for function \code{nrp} in Package `asuR'
}
\section{Methods}{
\describe{
\item{mymodel = "ANY"}{ generic function }

\item{mymodel = "lmer"}{ see the documentation of nrp.lmer} 
}}
\keyword{methods}
