\name{parseFormula-methods}
\docType{methods}
\alias{parseFormula-methods}
\alias{parseFormula,ANY-method}
\alias{parseFormula,lm-method}
\alias{parseFormula,glm-method}
\alias{parseFormula,lmer-method}
\title{ Methods for Function parseFormula in Package `asuR' }
\description{
  Methods for function \code{parseFormula} in Package `asuR' 
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{ generic function }

\item{mymodel = "lm"}{ see its man page }

\item{mymodel = "glm"}{ see its man page }

\item{mymodel = "lmer"}{ see its man page }
}}
\keyword{methods}
