\name{parseFormula.lmer}
\alias{parseFormula.lmer}
\title{providing names and data from model objects}
\description{
internal use only
}
\usage{
\method{parseFormula}{lmer}(mymodel, ...)
}
\arguments{
  \item{mymodel}{ a fitted model object }
  \item{...}{further arguments}
}
\details{
}
\value{
 
}
\author{\email{thomas.fabbro@unibas.ch}}
\note{under construction}
\seealso{\code{\link{inspect}}}
\examples{
##
}
\keyword{models}