\name{rfp-methods}
\docType{methods}
\alias{rfp-methods}
\alias{rfp,ANY-method}
\alias{rfp,lmer-method}
\title{Methods for Function rfp in Package `asuR'}
\description{
 Methods for function \code{rfp} in Package `asuR'
}
\section{Methods}{
\describe{
\item{mymodel = "ANY"}{ generic function }

\item{mymodel = "lmer"}{ see the documentation of rfp.lmer} 
}}
\keyword{methods}
