\name{rgp-methods}
\docType{methods}
\alias{rgp-methods}
\alias{rgp,ANY-method}
\alias{rgp,lmer-method}
\title{Methods for Function rgp in Package `asuR'}
\description{
 Methods for function \code{rgp} in Package `asuR'
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{ generic function }

\item{mymodel = "lmer"}{ see the documentation of rgp.lmer} 
}}
\keyword{methods}
