\name{rpp-methods}
\docType{methods}
\alias{rpp-methods}
\alias{rpp,ANY-method}
\alias{rpp,lm-method}
\alias{rpp,glm-method}
\title{Methods for Function rpp in Package `asuR'}
\description{
Creates a plot of partial residuals against all predictors.
}
\section{Methods}{
\describe{

\item{mymodel = "ANY"}{Generic function}

\item{mymodel = "lm"}{Generates a plot of an object of class "lm"}

\item{mymodel = "glm"}{Generates a plot of an object of class "glm"}
}}
\keyword{methods}
