\name{rpp}
\alias{rpp}
\alias{partialResidual.eachPredictor}
\title{partial residual vs. each predictor plot}
\description{
  There is a separate help page for every generic.
}
\usage{
rpp(mymodel, id = c("all", "none"), ...)
}
\arguments{
  \item{mymodel}{an fitted model object}
  \item{id}{a character string or numeric value; indicating in which
    panel it should be possible to interactively identify points}
  \item{...}{further arguments}
}
\details{
 There is a separate help page for every generic.
}
\value{
  A plot and an invisible vector of row indices (in the original data) of identified values.
}
\author{\email{thomas.fabbro@unibas.ch}}
\seealso{\code{\link{rpp.glm}}, \code{\link{inspect}}}
\examples{
##
}
\keyword{models}
