\name{length-methods}
\docType{methods}
\alias{length-methods}
\alias{length,ANY-method}
\alias{length,pDescList-method}
\title{Length of a pDescList-object}
\description{
Returns the length of a pDescList-object
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{see, definition of methods}

\item{\code{signature(x = "pDescList")}}{to be done}
\item{\code{signature(x = "Formula")}}{
sdf
}

\item{\code{signature(x = "pDescList")}}{
extract a pDesc object from a pDescList
}
}}
\keyword{methods}
\keyword{internal}

